Presentations
Applying various techniques to this security device is one of the emerging AI trends. our idea is about how to develop a weight recognition-based identification system that will include a buzzer system and lots of features.
In order for the system to function, it’s necessary to implement three steps. First, it must detect a weight. Then, it must recognize that weight of person nearly instantaneously. Finally, it must take whatever further action is required, such as if the weight that will be put on the window will be more than acquired weight then it will start buzzing because no person comes from window.