Screenshots and processing
In this project, we have a BLE and a sensor to collect data from the environment. The camera will be used for capturing the weight from the environment and sensor will be used for getting depth information. Data collected from these units will be sent to the server for further processing. Received weights will be processed on the server using processing techniques. This project also enables remote access of a relative or a friend of the user to the device. In order to assist the user, an alert sound will be made so that sound can be reached to relative or a friend and they will be able to help. To get the knowledge of user's environment direct video stream module will be used. 

Product Features

This device will have various features like:

*weight recognition*

*Buzzer alarm*

*Memory card to store information*

*Alert sound making technique*